Added ELEMENT_TYPE__EQUALS_FUNCTION parameter to macro ARRAY__DEFINE. This is a user-supplied equality comparison function, with signature bool ( *equals_function )( ELEMENT_TYPE const *first, ELEMENT_TYPE const *second ).
Added method array__equals. This is a method which compares two array structures for equality. That is, two arrays are equal if they have the same element type and length, and they both contain the same elements (ordering is significant). Returns true if the two arrays are equal, and false otherwise.

Added test array__char__equals.